requirements.txt
google-genai
groq
pyspark
requests
python-dotenv
pandas